This parses stuff now.
