+++ Gemfile
@@ -1,0 +1,3 @@
+:source rugbygems
+require 'couchrest'
+require 'json'
