+++ lib/aggregator.rb
@@ -1,0 +1,7 @@
+$: << File.dirname(__FILE__)
+
+require 'json'
+# require 'couchrest'
+require 'aggregator/entry'
+require 'aggregator/couchdb'
+require 'net/http'
